# src/api/docs/swagger.yaml
basePath: /api
definitions:
...
  handlers.ErrorResponse:
    properties:
      error:
        example: Something went wrong
        type: string
    type: object
...
  handlers.WishlistSearchAlertCandidateListResponse:
    properties:
      candidates:
        items:
          $ref: '#/definitions/models.AlertCandidate'
        type: array
      limit:
        example: 20
        type: integer
      page:
        example: 1
        type: integer
      total:
        example: 1
        type: integer
    type: object
  handlers.WishlistSearchAlertConvertRequest:
    properties:
      acknowledgeDuplicateWarning:
        example: false
        type: boolean
      coin:
        $ref: '#/definitions/models.Coin'
    type: object
  handlers.WishlistSearchAlertConvertResponse:
    properties:
      candidate:
        $ref: '#/definitions/models.AlertCandidate'
      coin:
        $ref: '#/definitions/models.Coin'
      warnings:
        items:
          type: string
        type: array
    type: object
  handlers.WishlistSearchAlertCriteriaAdjustRequest:
    properties:
      candidateIds:
        items:
          type: integer
        type: array
      criteria:
        $ref: '#/definitions/handlers.WishlistSearchAlertCriteriaRequest'
    type: object
...
  handlers.WishlistSearchAlertDismissRequest:
    properties:
      notes:
        example: Too expensive with shipping
        type: string
      reason:
        example: price_too_high
        type: string
    type: object
  handlers.WishlistSearchAlertListResponse:
    properties:
      alerts:
        items:
          $ref: '#/definitions/models.WishlistSearchAlert'
        type: array
      limit:
        example: 20
        type: integer
      page:
        example: 1
        type: integer
      total:
        example: 1
        type: integer
    type: object
  handlers.WishlistSearchAlertRequest:
    properties:
      cadence:
        example: manual
        type: string
      criteria:
        $ref: '#/definitions/handlers.WishlistSearchAlertCriteriaRequest'
      isActive:
        example: true
        type: boolean
      name:
        example: Domitian denarius under $300
        type: string
    type: object
...
  handlers.WishlistSearchAlertRunRequest:
    properties:
      maxCandidates:
        example: 20
        type: integer
    type: object
  handlers.WishlistSearchAlertRunResponse:
    properties:
      alertId:
        example: 1
        type: integer
      candidates:
        items:
          $ref: '#/definitions/models.AlertCandidate'
        type: array
      completedAt:
        example: "2026-06-29T17:00:10Z"
        type: string
      dismissedCount:
        example: 0
        type: integer
      duplicateCount:
        example: 3
        type: integer
      newCount:
        example: 5
        type: integer
      partialWarnings:
        items:
          type: string
        type: array
      rateLimitStatus:
        example: ok
        type: string
      resultCount:
        example: 8
        type: integer
      runId:
        example: 42
        type: integer
      startedAt:
        example: "2026-06-29T17:00:00Z"
        type: string
      status:
        allOf:
        - $ref: '#/definitions/models.AlertRunStatus'
        example: completed
    type: object
...
  models.AlertCandidate:
    properties:
      alertId:
        type: integer
      canonicalSourceUrl:
        type: string
      convertedCoinId:
        type: integer
      createdAt:
        type: string
      dismissalReason:
        type: string
      duplicateKey:
        type: string
      duplicateOfCandidateId:
        type: integer
      fields:
        $ref: '#/definitions/models.StringMap'
      firstSeenAt:
        type: string
      id:
        type: integer
      lastSeenAt:
        type: string
      lifecycleState:
        $ref: '#/definitions/models.AlertCandidateState'
      matchingWishlistCoinId:
        type: integer
      normalizedTitle:
        type: string
      observedCurrency:
        type: string
      observedPrice:
        type: number
      provenance:
        items:
          $ref: '#/definitions/models.CandidateProvenance'
        type: array
      provenanceStatus:
        $ref: '#/definitions/models.CandidateProvenanceStatus'
      reasonForMatch:
        type: string
      runId:
        type: integer
      sourceName:
        type: string
      sourceUrl:
        type: string
      title:
        type: string
      updatedAt:
        type: string
      userId:
        type: integer
    type: object
...
  models.StringMap:
    additionalProperties:
      type: string
    type: object
...
  models.WishlistSearchAlert:
    properties:
      cadence:
        $ref: '#/definitions/models.WishlistAlertCadence'
      coinType:
        type: string
      createdAt:
        type: string
      currency:
        type: string
      dateFrom:
        type: integer
      dateTo:
        type: integer
      dealerPreference:
        type: string
      gradeOrCondition:
        type: string
      id:
        type: integer
      isActive:
        type: boolean
      keywords:
        type: string
      lastRunAt:
        type: string
      material:
        type: string
      mint:
        type: string
      name:
        type: string
      notes:
        type: string
      priceMax:
        type: number
      priceMin:
        type: number
      rulerOrIssuer:
        type: string
      sourceFilters:
        items:
          type: string
        type: array
      updatedAt:
        type: string
      userId:
        type: integer
    type: object
...
info:
  contact: {}
  description: REST API for managing an ancient coin collection. Supports coin CRUD,
    image uploads, AI-powered analysis via Ollama, user management, and admin features.
  title: Ancient Coins API
  version: "1.0"
paths:
...
  /wishlist/search-alerts:
    get:
      description: Lists acquisition discovery search alerts. This is separate from
        wishlist availability checking.
      parameters:
      - description: Filter by active state
        in: query
        name: active
        type: boolean
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.WishlistSearchAlertListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List wishlist search alerts
      tags:
      - Wishlist Search Alerts
    post:
      consumes:
      - application/json
      description: Creates a saved discovery alert without creating wishlist items
        or availability rows.
      parameters:
      - description: Alert
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.WishlistSearchAlertRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.WishlistSearchAlert'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create wishlist search alert
      tags:
      - Wishlist Search Alerts
  /wishlist/search-alerts/{alertId}:
    delete:
      description: Soft-deletes a saved discovery alert while preserving history.
      parameters:
      - description: Alert ID
        in: path
        name: alertId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete wishlist search alert
      tags:
      - Wishlist Search Alerts
    get:
      description: Gets one acquisition discovery alert owned by the authenticated
        collector.
      parameters:
      - description: Alert ID
        in: path
        name: alertId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.WishlistSearchAlert'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get wishlist search alert
      tags:
      - Wishlist Search Alerts
    put:
      consumes:
      - application/json
      description: Updates saved discovery criteria/cadence/active metadata without
        changing availability state.
      parameters:
      - description: Alert ID
        in: path
        name: alertId
        required: true
        type: integer
      - description: Alert
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.WishlistSearchAlertRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.WishlistSearchAlert'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update wishlist search alert
      tags:
      - Wishlist Search Alerts
  /wishlist/search-alerts/{alertId}/candidates:
    get:
      parameters:
      - description: Alert ID
        in: path
        name: alertId
        required: true
        type: integer
      - description: Candidate state
        in: query
        name: state
        type: string
      - description: Provenance status
        in: query
        name: provenanceStatus
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.WishlistSearchAlertCandidateListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List wishlist search alert candidates
      tags:
      - Wishlist Search Alerts
  /wishlist/search-alerts/{alertId}/candidates/{candidateId}/convert:
    post:
      consumes:
      - application/json
      parameters:
      - description: Alert ID
        in: path
        name: alertId
        required: true
        type: integer
      - description: Candidate ID
        in: path
        name: candidateId
        required: true
        type: integer
      - description: Convert request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.WishlistSearchAlertConvertRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.WishlistSearchAlertConvertResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Convert wishlist search alert candidate
      tags:
      - Wishlist Search Alerts
  /wishlist/search-alerts/{alertId}/candidates/{candidateId}/dismiss:
    post:
      consumes:
      - application/json
      parameters:
      - description: Alert ID
        in: path
        name: alertId
        required: true
        type: integer
      - description: Candidate ID
        in: path
        name: candidateId
        required: true
        type: integer
      - description: Dismiss request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.WishlistSearchAlertDismissRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AlertCandidate'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Dismiss wishlist search alert candidate
      tags:
      - Wishlist Search Alerts
  /wishlist/search-alerts/{alertId}/candidates/{candidateId}/restore:
    post:
      parameters:
      - description: Alert ID
        in: path
        name: alertId
        required: true
        type: integer
      - description: Candidate ID
        in: path
        name: candidateId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AlertCandidate'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Restore wishlist search alert candidate
      tags:
      - Wishlist Search Alerts
  /wishlist/search-alerts/{alertId}/criteria-adjustments:
    post:
      consumes:
      - application/json
      parameters:
      - description: Alert ID
        in: path
        name: alertId
        required: true
        type: integer
      - description: Criteria adjustment
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.WishlistSearchAlertCriteriaAdjustRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.WishlistSearchAlert'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Adjust wishlist search alert criteria
      tags:
      - Wishlist Search Alerts
  /wishlist/search-alerts/{alertId}/run:
    post:
      consumes:
      - application/json
      description: Manually runs a saved discovery alert and stores candidates separately
        from availability checks.
      parameters:
      - description: Alert ID
        in: path
        name: alertId
        required: true
        type: integer
      - description: Run request
        in: body
        name: body
        schema:
          $ref: '#/definitions/handlers.WishlistSearchAlertRunRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.WishlistSearchAlertRunResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Run wishlist search alert
      tags:
      - Wishlist Search Alerts
  /wishlist/search-alerts/{alertId}/runs:
    get:
      parameters:
      - description: Alert ID
        in: path
        name: alertId
        required: true
        type: integer
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.WishlistSearchAlertRunListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List wishlist search alert runs
      tags:
      - Wishlist Search Alerts
  /wishlist/search-alerts/{alertId}/runs/{runId}:
    get:
      parameters:
      - description: Alert ID
        in: path
        name: alertId
        required: true
        type: integer
      - description: Run ID
        in: path
        name: runId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AlertRun'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get wishlist search alert run
      tags:
      - Wishlist Search Alerts
securityDefinitions:
  ApiKeyAuth:
    description: Enter your API key, e.g. "ak_a1b2c3d4..."
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: Enter your JWT token with the Bearer prefix, e.g. "Bearer eyJhbGci..."
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"